Fixed all issues. No cors policy error. Changed most of the database and switched from hibernate to flyway.

+++ src/main/java/nl/fontys/s3/erp/business/impl/ProductsImpl/DeleteProductUseCaseImpl.java
@@ -3,6 +3,8 @@ package nl.fontys.s3.erp.business.impl.ProductsImpl;
 
 import lombok.AllArgsConstructor;
 import nl.fontys.s3.erp.business.ProductsUseCases.DeleteProductUseCase;
+import nl.fontys.s3.erp.business.exceptions.PermissionDenied;
+import nl.fontys.s3.erp.configuration.security.token.AccessToken;
 import nl.fontys.s3.erp.persistence.ProductRepository;
 import org.springframework.stereotype.Service;
 
@@ -10,9 +12,13 @@ import org.springframework.stereotype.Service;
 @AllArgsConstructor
 public class DeleteProductUseCaseImpl implements DeleteProductUseCase {
     private final ProductRepository productRepository;
+    private final AccessToken accessToken;
 
     @Override
     public void deleteProduct(long id){
+        if(accessToken.getDepartments() == null || !accessToken.getDepartments().contains("PRODUCT")) {
+            throw new PermissionDenied("delete a product.");
+        }
         productRepository.deleteById(id);
     }
 }
